"feat: add condition to delete quote"

+++ src/usecases/delete-quote-usecase.ts
@@ -1,9 +1,20 @@
 import { QuoteRepository } from '../gateways';
-import { Either } from '../shared';
+import { Either, left } from '../shared';
+import { QuoteNotExistError } from './errors';
+import { QuoteNotBelongsMeError } from './errors/quote-not-belongs-me';
 
 export class DeleteQuoteUseCase {
   constructor(private readonly quoteRepository: QuoteRepository) {}
-  async execute(quoteId: string): Promise<Either<Error, string>> {
+  async execute(
+    quoteId: string,
+    currentUserId: string
+  ): Promise<Either<Error, string>> {
+    const quote = await this.quoteRepository.findById(quoteId);
+    if (!quote) return left(new QuoteNotExistError());
+
+    if (!quote.canBeDeleted(currentUserId))
+      return left(new QuoteNotBelongsMeError());
+
     return await this.quoteRepository.delete(quoteId);
   }
 }
